requirements.txt
django>=5.0 
djangorestframework>=3.15
langchain>=0.3.27
langchain-community>=0.3.27
langchain_huggingface>=0.3.1
sentence-transformers>=3.0.1
faiss-cpu>=1.8.0
pymupdf>=1.24.9 
pypdf>=5.0.0
sympy>=1.13
langchain-ollama
python-decouple==3.8
